pyproject.toml
[tool.poetry]
name = "twitter_blocklist"
version = "0.2.1"
description = "Export and import Twitter blocklists"
authors = ["Andrea Zonca <code@andreazonca.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/zonca/twitter_blocklist"

[tool.poetry.dependencies]
python = "^3.7"
python-twitter = "^3.5"
toml = "^0.10.1"
click = "^7.1.2"
progressbar2 = "^3.51.3"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
twitter_blocklist = "twitter_blocklist.console:main"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

